Fold all relations into a Relation type
`is` and `within` are both legacy forms. In terms of semantics, `n is 1`
is exactly equivalent to `n = 1`. The difference is only in the grammar;
the `is` form only allows a number on the right, not `n is 1..9`.

Similarly, `within` has the same grammar as `=` and `in`, and only
changes the way ranges are processed.